Enum spatial::tree::mbr::MbrNode

pub enum MbrNode<P, DIM, LG, T> where DIM: ArrayLength<P> + ArrayLength<(P, P)> {
    Level {
        mbr: Rect<P, DIM>,
        children: Vec<MbrNode<P, DIM, LG, T>>,
    },
    Leaves {
        mbr: Rect<P, DIM>,
        children: Vec<RwLock<MbrLeaf<P, DIM, LG, T>>>,
    },
}

Level node of a tree. Either contains other levels or leaves

Variants

Level

Contains only other levels

Fields

mbr

children

Leaves

Contains only leaves

Fields

mbr

children

Methods

impl<P, DIM, LG, T> MbrNode<P, DIM, LG, T> where P: Float + Signed + Bounded + MulAssign + AddAssign + ToPrimitive + FromPrimitive + Copy + Debug + Default, DIM: ArrayLength<P> + ArrayLength<(P, P)>, LG: MbrLeafGeometry<P, DIM>

fn new_leaves() -> MbrNode<P, DIM, LG, T>

Create an empty leaf level

fn new_no_alloc() -> MbrNode<P, DIM, LG, T>

Create an empty leaf level with no capacity for leaves. Only used for passing ownership of root into the index functions

fn has_leaves(&self) -> bool

Does the level point to leaves?

fn has_levels(&self) -> bool

Does the level point to other levels?

fn mbr(&self) -> &Rect<P, DIM>

Borrow the level's minimum bounding rectangle

fn mbr_mut(&mut self) -> &mut Rect<P, DIM>

Mutably borrow the level's minimum bounding rectangle

fn len(&self) -> usize

Number of level's children

fn is_empty(&self) -> bool

Does the level have children?